implement SortButton component and update project details tabs
