+++ frontend/packages/app/src/pages/project-details/tabs/index.tsx
@@ -11,8 +11,8 @@ export const TAB_KEYS = [
   "overview",
   "calendar",
   "tracking",
-  "notes",
   "risks",
+  "notes",
   "email",
   "to-do",
   "feedback",
